Add signposts for evaluating UA JavaScript in WebProcess https://bugs.webkit.org/show_bug.cgi?id=302350 rdar://164507890
Reviewed by NOBODY (OOPS!).

This is similar to 302769@main, except that these signposts are in WebProcess rather than in
UIProcess. This is useful for understanding the actual wall clock time associated with JS execution
(the previous signposts in UIProcess also capture IPC latency, which isn't what we care about when
optimizing benchmarks like Speedometer).

* Source/WTF/wtf/SystemTracing.h:
* Source/WebKit/WebProcess/WebPage/WebPage.cpp:
(WebKit::WebPage::runJavaScript):